pyproject.toml
[project]
name = "hinglish-collection-agent"
version = "0.1.0"
description = "Real-time Hinglish voice agent for post-default debt collection."
requires-python = ">=3.11,<3.13"
dependencies = [
    "pipecat-ai[local,openai,runner,silero,webrtc]>=0.0.47",
    "faster-whisper>=1.0.3",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "soundfile>=0.12.1",
    "sounddevice>=0.4.7",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "pydantic>=2.7.0",
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
]

[project.optional-dependencies]
eval = [
    "deepgram-sdk>=3.5.0",
    "jiwer>=3.0.4",
    "matplotlib>=3.9.0",
    "pandas>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stt", "tts", "nlp", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"
